+++ next/api/src/json/file.ts
@@ -1,4 +1,4 @@
-import { File } from '../model2/File';
+import { File } from '../model/File';
 
 export class FileResponse {
   constructor(readonly file: File) {}
