+++ next/api/src/json/user.ts
@@ -2,7 +2,7 @@ import crypto from 'crypto';
 import LRU from 'quick-lru';
 
 import { config } from '../config';
-import { User } from '../model2/User';
+import { User } from '../model/User';
 
 class GravatarUrlManager {
   static cache = new LRU<string, string>({ maxSize: 100_000 });
