+++ next/api/src/router/category.ts
@@ -3,8 +3,8 @@ import Router from '@koa/router';
 
 import * as yup from '../utils/yup';
 import { auth } from '../middleware/auth';
-import { Category, CategoryManager } from '../model2/Category';
-import { TicketForm } from '../model2/TicketForm';
+import { Category, CategoryManager } from '../model/Category';
+import { TicketForm } from '../model/TicketForm';
 import { CategoryJson, CategoryFieldResponse } from '../json/category';
 
 const router = new Router().use(auth);
